Update is_multiple.py
Added protection for divide by zero on the second entry m. split main to add a getInput function.

Entry of 0 for n is still allowed as n = mi which could be 0 = m*0 and n could be = i. Not sure I like that and it would seem better to not allow, but seems the definition supports it so trying to follow the definition. Glad to update and prevent 0 if that is not the intent. Please let me know.