Update punishment.py
Trying to update pull request and review. I edited the print string formatting to use the short cut method and left the long version in a comment so it can be discussed in the video. Not sure if I am getting the updates properly entered in git, so will see how this goes.