use of continue
I commented out 2 uses of continue for the excepts as they were not needed and retained continue for the handling of divide by zero. I also added a clarification on the print statement for divide by zero. The last submission worked and was correct but had the two uncecessary lines so this is a light clean up and provides a more accurate picture of when and when not to use continue.