wk 4 linked stack, linked queue, and linked deque
First and Final submission in git Hub. This is an update to canvas and these updates will be posted as fnal in canvas with updated videos. This updates the .size method for stack and queue. Not that my program uses them as methods, but I wanted the size function of all three to work the same. Most updates are grammar beyond that in comments. Size will report if 0, return 0, vice none or raising an error or Empty. This is greatly explained in the Deque video.